Move apply_diff to a tool file
